All blog
Se actualizo el blog del proyecto

+++ resources/views/cajas/index.blade.php
@@ -1,27 +1,36 @@
-<a href="/cajas/create">Create</a><br>
-<table class="table">
+@extends('/layoutProyect')
+
+ @section('indexCajas')
+<h3 class="m3">Cajas</h3>
+<table class="table centered striped">
 	<tr>
-		<th>ID</th>
-		<th>Description</th>
-		<th>Cliente</th>
+		<th class="center">ID</th>
+		<th class="center">Descripcion</th>
+		<th class="center">Cliente(s)</th>
+    <th class="center">Acciones</th>
 	</tr>
 	<tr>
 		@foreach($cajas as $caja)
 		<td>{{$caja->id}}</td>
 		<td>{{$caja->description}}</td>
-		<td>{{$caja->cliente_id}}</td>
+		<td>{{$caja->cdesc}} {{$caja->cdescc}}</td>
 
 		<td>
-			{!! Form::open(['url' => '/cajas/'.$caja->id,'method'=>'delete']) !!}
-			
-			<a class="waves-effect waves-light btn-small" href="{!! '/cajas/'.$caja->id.'/edit' !!}">Editar</a>
+			{!! Form::open(['url' => '/cajas/'.$caja->id,'method'=>'delete','onsubmit'=>'return ConfirmDelete()']) !!}
 
-			{!! Form::button('Delete',['type'=>'submit','class'=>'waves-effect waves-light btn-small']) !!}
+			<a class="waves-effect waves-light btn light-blue darken-3" href="{!! '/cajas/'.$caja->id.'/edit' !!}"><i class="material-icons right">edit</i>Editar</a>
+
+			{!! Form::button('<i class="material-icons right">delete</i>Borrar',['type'=>'submit','class'=>'waves-effect waves-light btn red']) !!}
 
 			{!! Form::close()!!}
 		</td>
-		
+
 	</tr>
 	@endforeach
-
-</table>
+</table></br>
+<div class="row">
+<div class="col s5 offset-s9">
+<a class="waves-effect waves-light btn light-blue darken-1" href="/cajas/create"><i class="material-icons right">add</i>agregar caja</a><br>
+</div>
+</div>
+@endsection
